+++ lua/miriani/scripts/combat.lua
@@ -50,6 +50,7 @@ ImportXML([=[
    mplay("ship/combat/hit/youHit", "ship", nil, nil, nil, nil, nil, nil, volume_offset)
    print_color({"Hit by %1", "combat"})
    channel("combat", "Hit by %1", {"combat"})
+   update_artifact_hunting_activity()
   </send>
   </trigger>